Move view to correct module
